+++ R/param_num_attn_blocks.R
@@ -5,7 +5,7 @@
 #' @inheritParams Laplace
 #'
 #' @details
-#' Used as a tuning parameter for `tabular_auto_int()` in the `tdl` package
+#' Used as a tuning parameter for `tabular_auto_int()` in the \pkg{tabular} package
 #' when fit with the `brulee` engine. Each attention block consists of a
 #' multi-head attention layer followed by feed-forward layers.
 #'