+++ R/param_penalty_type.R
@@ -6,7 +6,7 @@
 #' @param values A character vector of possible values.
 #'
 #' @details
-#' Used as a tuning parameter for `tabular_rln()` in the `tdl` package when
+#' Used as a tuning parameter for `tabular_rln()` in the \pkg{tabular} package when
 #' fit with the `brulee` engine. L1 is recommended by the original paper.
 #'
 #' @examples