+++ R/param_resid_at.R
@@ -6,7 +6,7 @@
 #' @inheritParams Laplace
 #'
 #' @details
-#' Used as a tuning parameter for `tabular_resnet()` in the `tdl` package
+#' Used as a tuning parameter for `tabular_resnet()` in the \pkg{tabular} package
 #' when fit with the `brulee` engine. The upper bound depends on the number
 #' of hidden layers in the network and so is left as `unknown()` by default.
 #'