+++ NEWS.md
@@ -1,3 +1,5 @@
+# dials (development version)
+
 # dials 1.4.3
 
 ## New parameters