+++ NEWS.md
@@ -1,3 +1,5 @@
+# dials (development version)
+
 # dials 1.4.4
 
 * `attention_type()`, `dropout_hidden()`, `dropout_last()`, and `target_token()` are new parameters for the SAINT model in the brulee package (#452).